refactor(svelte2tsx): extract runes detection into script/runes.rs
Pure move: no logic changes, only `use` additions and `pub(super)` visibility
adjustments. The corresponding unit tests move with the code.
